Remove one copy of duplication from IPC types
