+++ desktop/src/services/imageProcessor.ts
@@ -8,7 +8,7 @@ import { CustomErrors } from "../constants/errors";
 import { writeStream } from "../main/fs";
 import { isDev } from "../main/general";
 import { logError, logErrorSentry } from "../main/log";
-import { ElectronFile } from "../types";
+import { ElectronFile } from "../types/ipc";
 import { isPlatform } from "../utils/common/platform";
 import { generateTempFilePath } from "../utils/temp";
 import { deleteTempFile } from "./ffmpeg";
