+++ desktop/src/services/upload.ts
@@ -1,7 +1,8 @@
 import StreamZip from "node-stream-zip";
 import path from "path";
 import { uploadStatusStore } from "../stores/upload.store";
-import { ElectronFile, FILE_PATH_KEYS, FILE_PATH_TYPE } from "../types";
+import { ElectronFile, FILE_PATH_TYPE } from "../types/ipc";
+import { FILE_PATH_KEYS } from "../types/main";
 import { getValidPaths, getZipFileStream } from "./fs";
 
 export const getSavedFilePaths = (type: FILE_PATH_TYPE) => {
