--- desktop/src/types/autoLauncher.ts
@@ -1,5 +1,0 @@
-export interface AutoLauncherClient {
-    isEnabled: () => Promise<boolean>;
-    toggleAutoLaunch: () => Promise<void>;
-    wasAutoLaunched: () => Promise<boolean>;
-}
